UPSTREAM: power: supply: sbs-battery: Prevent CAPACITY_MODE races
A subset of smart battery commands return charge or energy depending on
the CAPACITY_MODE bit setting of BatteryMode(). In order to
unambiguously read a charge or energy value, it is necessary to ensure
that CAPACITY_MODE is set as desired, and not changed for the duration
of the attribute read.

Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.co.uk>
(cherry picked from commit fe8a6534396807b8a0863cd146b72c0bce0b0c88)

BUG=chromium:848112
TEST=kevin and scarlet battery

Change-Id: I214d2d342ca34ea0ec0add4e90a6e72a3f32e917
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1081802
Reviewed-by: Guenter Roeck <groeck@chromium.org>