UPSTREAM: r8152: add byte_enable for ocp_read_word function
Add byte_enable for ocp_read_word() to replace reading 4
bytes data with reading the desired 2 bytes data.

This is used to avoid the issue which is described in
commit b4d99def0938 ("r8152: remove sram_read"). The
original method always reads 4 bytes data, and it may
have problem when reading the PHY registers.

The new method is supported since RTL8153B, but it
doesn't influence the previous chips. The bits of the
byte_enable for the previous chips are the reserved
bits, and the hw would ignore them.

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit d8fbd27469fc02049c674de296a3263bef089131)

BUG=chromium:768542
TEST=netperf, repeated reboot, and network_EthernetStressPlug on coral

Change-Id: Ia62c667c70c2766db02d34a874765bf56dfe80f5
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/865148
Reviewed-by: Guenter Roeck <groeck@chromium.org>