UPSTREAM: drm/i915: Store a persistent reference for an object in the execbuffer cache
If we take a reference to the object/vma when it is first used in an
execbuf, we can keep that reference until the object's file-local handle
is closed. Thereby saving a frequent ref/unref pair.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
(cherry picked from commit dade2a6165fd7bd825958e9b0ad04678f53f0f60)
Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>

BUG=None
TEST=CNL boots to graphics

Change-Id: I2868fa9e68b03815ddf8d8770a0b9f3450baca27