BACKPORT: drm/i915: Allow execbuffer to use the first object as the batch
Currently, the last object in the execlist is the always the batch.
However, when building the batch buffer we often know the batch object
first and if we can use the first slot in the execlist we can emit
relocation instructions relative to it immediately and avoid a separate
pass to adjust the relocations to point to the last execlist slot.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
(cherry picked from commit 1a71cf2fa646799d4397a49b223549d8617fece0)
Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>

BUG=None
TEST=CNL boots to graphics

Conflicts:
	Nathan - context delta:
	include/uapi/drm/i915_drm.h

Change-Id: Ic43fdca0b65a0f967a08975032253d2128dd591f