UPSTREAM: ASoC: Intel: bxt: Move codec sysclk config to codec_init function
The MCLK for DA7219 does not change in this platform, but is
currently being configured everytime as part of the platform_clock
event handler for DAPM. The upshot of this is that we have
unnecessary calls to this function, and it also means that if
a stream hasn't yet been started, DA7219 driver does not have the
correct MCLK rates programmed and so the HP detection feature does
not operate as expected.

This patch rectifies this issue by moving the sysclk call to
codec_init function so it's only called once at initialisation.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Acked-by: Sathyanarayana Nujella <sathyanarayana.nujella@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git/
tag: for-next
(cherry picked from commit fd0f237572ada2b18fe1e8684ae9fa131b577a17)

BUG=b:36468989,b:37455154
TEST=cross checked HP detection, Audio scenarios works
Signed-off-by: Sathyanarayana Nujella <sathyanarayana.nujella@intel.com>

Change-Id: Iaf527ae7620d050a47a763b3cd88b5cd75c301fe
Signed-off-by: Sathyanarayana Nujella <sathyanarayana.nujella@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/557977
Tested-by: Kuen Liu <kuen.liu@quanta.corp-partner.google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>