CHROMIUM: ARM: errata: add support for A12/A17 errata CR711784
This adds a code for turning on chicken bit 11, which appears to avoid
a potential CPU deadlock that could occur.  The exact set of
instruction needed to trigger this errata is not totaly known yet, but
we have a relatively high level of confidence that the problem is
fixed by setting chicken bit 11.

All details are in http://crbug.com/711784

This errata does not yet have a name, so I have tagged it
ERRATA_CR711784.  It will be up to ARM to upstream this errata once
they have fully investigated it.

BUG=chromium:711784
TEST=youtube clicking, as described in the BUG
TEST=for i in $(seq 0 3); do taskset -c $i cat /sys/*/*/*or_debug; done

Change-Id: I0aeac31882c9b39407d8c96333a237797a9f7000
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/541537
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>