UPSTREAM: drm/i915: Cancel pending execlist tasklet upon wedging
Highly unlikely, but if the stop_machine() did suspend the tasklet, we
want to make sure that when it wakes it finds there is nothing to do.
Otherwise, it will loudly complain that the ELSP port tracking no longer
matches the hardware, and we will be mightly confused.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20170621124804.4529-1-chris@chris-wilson.co.uk
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
(cherry picked from commit 4ee056f41807858b2eae263e74ae8b81800c0337)
Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>

BUG=None
TEST=CNL boots to graphics

Change-Id: I63c86f19208efd76943ba420afb9b11a4cdf910c