HACK: CHROMIUM: x86/mm: don't check %sp when faulting page immediately before stack
Prior to 1be7107fbe18 ("mm: larger stack guard gap, between vmas"), the
stack guard page was included in the stack's VMA.  This implementation
detail created a subtle corner case in __do_page_fault(), where the
branch

	if (likely(vma->vm_start <= address))
		goto good_area;

would skip over the %sp check when `address' was inside the stack guard
page.  This allowed userspace to expand the stack without altering %sp
by repeatedly touching the page just below the "advertised" start of the
stack VMA.

Although this corner case probably isn't intended, it appears to have
been in the kernel for years, and is used by real software in the wild.
For example the Android Runtime's Thread::InstallImplicitProtection()
routine has relied on this behavior to expand the main thread's stack,
and crashes on x86 kernels with 1be7107fbe18 applied.

NOTE: Marked as HACK since it effectively re-introduces a bug into the
kernel which is exploited by zygote in Android. Fixing that bug results
in a zygote crash in art::Thread::InstallImplicitProtection(). If and when
this problem will be fixed is currently unknown. If possible, this patch
should be dropped at a later time or, if applicable, be replaced with an
upstream patch.

BUG=chromium:726072, chromium:724093, b:62952017, chromium:737932
TEST=Run ARC++

Change-Id: I8d2f6d3e598e99b2605a9caf8e5c6c166de88e0f
Fixes: 1be7107fbe18 ("mm: larger stack guard gap, between vmas")
Signed-off-by: Greg Hackmann <ghackmann@google.com>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/559932
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit 3c2f4706a9304f9723a85c935c540eaa088f5a43)
Reviewed-on: https://chromium-review.googlesource.com/562439