CHROMIUM: iwl7000: cfg80211: Add support for configuring NAN committed DW
Add support for configuring the NAN device committed Discovery Window
(DW) period, which allows the NAN device to not be present in all DWs.
Allowing the local device to not be present in all the DWs enables the
local device to perform operations on other channels during the
non-committed DWs, e.g., data path etc.

A peer device that desires to initiate communication with the local
device should do so in the committed DW of the local device.

ticket=none

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: 1fc3382ee88a0b6c4565dbe05d12f94e7e473cc7