CHROMIUM: tpm: ignore failed selftest in probe
If a tpm is in failed self-test mode, ConstinueSelfTest
command results in TPM_FAILEDSELFTEST error, and probe
fails. The tpm device is not created, and that prevents
the OS from attempting any further recover operations
with the tpm. Instead, ignore the negative result of the
self test, and create the device - the chip is out there,
it's just in the failed state.

BUG=none
TEST=put tpm in a failed selftest mode, boot, check that
     /dev/tpm0 is present, and GetCapability commands
     succeed.

Change-Id: I56b101850902d1983653bf29c8e0c662a5c7e8a0
Signed-off-by: Andrey Pronin <apronin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/567516
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>