CHROMIUM: iwl7000: mvm: set A-MPDU bit upon empty BA notification from FW
The bit was set only if there was at least one reclaimed frame in an
aggregation. It's important to set it also in the case that the whole
A-MPDU was lost, otherwise rate scaling statistics will not be
updated correctly. Thus, set it always in ba notification handler.

ticket=jira:WIFILNX-33

Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
iwl7000-tree: 8c471078eec8ca057f76a7d09e0cc063153acce1