FROMLIST: BACKPORT: intel_idle: Add S0ix validation
This adds validation of S0ix entry and enables it on Skylake. Using
the new tick_set_freeze_event function, we program the CPU to wake up
X seconds after entering freeze. After X seconds, it will wake the CPU
to check the S0ix residency counters and make sure we entered the
lowest power state for suspend-to-idle.

It exits freeze and reports an error to userspace when the SoC does
not enter S0ix on suspend-to-idle.

One example of a bug that can prevent a Skylake CPU from entering S0ix
(suspend-to-idle) is a leaked reference count to one of the i915 power
wells. The CPU will not be able to enter Package C10 and will
therefore use about 4x as much power for the entire system. The issue
is not specific to the i915 power wells though.

BUG=b:78841592
TEST=apply all patches and verify the S0ix check works (is invoked and
fails / passes as per the expectations)

(Applied from https://patchwork.kernel.org/patch/9831229/)

Change-Id: I99a99c02e66a5586b116d2fa2f0223f6980e7ebd
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Signed-off-by: Rajneesh Bhardwaj <rajneesh.bhardwaj@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1106584
Commit-Ready: Wenkai Du <wenkai.du@intel.corp-partner.google.com>
Tested-by: Wenkai Du <wenkai.du@intel.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>

[rebase419(groeck): Squash with
	FIXUP: FROMLIST: BACKPORT: intel_idle: Add S0ix validation]
Signed-off-by: Guenter Roeck <groeck@chromium.org>