CHROMIUM: drm/img-rogue: Always convert defreq_target to OPP frequency
The devfreq core expects a devfreq profile's target() callback to always
return with freq set to the current frequency - even if the function
does not actually change the frequency, it should still modify *freq
to match the current un-set value.

 * @target:
 *                      Before returning, target function should set
 *                      freq at the current frequency.

The devfeq core relies on the returned freq value being a correct OPP
frequency to perform its state transition statistics.

Note: This change was previously applied to the 1.7 driver (CL:505988),
but lost during the upgrade to 1.8.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:722329
TEST=http://webglsamples.org/aquarium/aquarium.html
  => dmesg does not contain
  "devfreq 13000000.mfgsys-gpu: Couldn't update frequency transition information."

Change-Id: I824744615711c9afbfcbb8edc6b56a7526f258da
Reviewed-on: https://chromium-review.googlesource.com/564877
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>