UPSTREAM: power: supply: sbs-battery: Add delay when changing capacity mode bit
At least with the Inspired Energy compatible batteries a delay is required
after setting the capacity mode bit from amp to watts or the reverse.
Setting the bit and then immediately pooling the status register results
in an unknown error being returned in the register. Add the delay results
in and ok status being return. This was also seen when reading the charge
and energy registers where the wrong value was returned for the requested
mode.

Signed-off-by: Phil Reid <preid@electromag.com.au>
Tested-by: Michael Heinemann <committed@heine.so>
Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.co.uk>
(cherry picked from commit adcf04c9f8fce54e65f11bae2fbd37c1282ef128)

BUG=chromium:848112
TEST=kevin and scarlet battery

Change-Id: I9ef34215bb70f809f09b2e737fcffca99379c990
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1081806
Reviewed-by: Guenter Roeck <groeck@chromium.org>