CHROMIUM: ext4: Hack to allow mounting partition with ext4crypto.
DO NOT MERGE TO NON-FACTORY BRANCHES.

In Chromium OS Factory flow, the manufacturing line may want to "reset"
a device by wiping stateful partition. However in order to get better
OOBE user experience, the CRX cache on stateful partition must be
preserved in wiping.

Starting from M59, kernel has enabled the ext4crypto feature so
partitions formatted by M59 would not be accessible by older kernels.
However, the only files we need (CRX cache) are not encrypted so it
should be totally fine to just ignore ext4crypto and mount partition
for reading those files since we will be formatting the partition
immediately after that.

This is a special hack only for factory branches otherwise we'll need to
either rebase kernel (and this is known to break on some older platforms
due to WiFi and BT stack changes) or cutting a new factory branch (which
will take factories weeks to months to re-stabilize).

BUG=b:62691173,b:63513991
TEST=Boots reset shim and mounts M59 stateful partition.

Change-Id: I2fcd1745d7de2543fdd18a19809ff8c9a8c73fd2
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/567028